backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.3
Pillow>=10.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
stripe>=7.0.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=21.2.0
qrcode[pil]>=7.4.2
pyotp>=2.9.0
pytest>=8.3.0
pytest-django>=4.8.0
